Fixed bug that caused the cloud DB to be cleared periodically, added login, role-based authorization and implemented improved logic for entity deletions

+++ routes/workspace.routes.mjs
@@ -1,7 +1,9 @@
 import express from "express";
 
-import {createWorkspace, deleteWorkspace, readWorkspace, updateWorkspace} from "../controllers/workspace_controller.mjs";
-import {addUserToWorkspace, deleteUserFromWorkspace, readWorkspaceForUser, updateUserInWorkspace} from "../controllers/workspaceUser_controller.mjs";
+import {createWorkspace, deleteWorkspace, readWorkspace, readAllWorkspaces, updateWorkspace} from "../controllers/workspace_controller.mjs";
+import {addUserToWorkspace, deleteUserFromWorkspace, readUserRoleInWorkspace, readWorkspaceCandidatesForDeletion, readWorkspaceForUser, updateUserInWorkspace} from "../controllers/workspaceUser_controller.mjs";
+import {deleteProject, readProjects} from "../controllers/project_controller.mjs";
+import {readUser} from "../controllers/user_controller.mjs";
 
 const router = express()
 
@@ -20,9 +22,9 @@ const router = express()
  *
  * Response Statuses
  * Success - 201 Created
- * Failure - 400 Bad Request - Missing a required attribute
+ * Failure - 400 Bad Request
  */
-router.post('/users/:user_id/workspaces/', function (req, res) {
+router.post('/users/:user_id/workspaces/', async function (req, res) {
     if (req.body.workspace_name == null) {
         res.status(400).json({
                 'Error':
@@ -31,13 +33,12 @@ router.post('/users/:user_id/workspaces/', function (req, res) {
             },
         )
     } else {
-        createWorkspace({
+        await createWorkspace({
             workspace_name: req.body.workspace_name,
             use_id: req.params.user_id
         })
-            .then(() => {
-                res.status(201).send()
-            })
+
+        res.status(201).send()
     }
 })
 
@@ -61,18 +62,23 @@ router.post('/users/:user_id/workspaces/', function (req, res) {
  *
  * Response Statuses
  * Success - 200 OK
+ * Failure - 403 Forbidden
  * Failure - 404 Not Found
  */
-router.get('/users/:user_id/workspaces/:workspace_id', function (req, res) {
-    // TODO - PASS JWT AS user_id FOR AUTHORIZATION???
-    readWorkspace(req.params.workspace_id)
-        .then(workspace => {
-            if (!workspace) {
-                res.status(404).json({'Error': 'No workspace with this workspace id exists'})
-            } else {
-                res.status(200).json(workspace)
-            }
-        })
+router.get('/users/:user_id/workspaces/:workspace_id', async function (req, res) {
+    const workspace = await readWorkspace(req.params.workspace_id)
+
+    if (workspace) {
+        const userRole = await readUserRoleInWorkspace(req.params.user_id, req.params.workspace_id)
+
+        if (userRole) {
+            res.status(200).json(workspace)
+        } else {
+            res.status(403).send()
+        }
+    } else {
+        res.status(404).json({'Error': 'No workspace with this workspace id exists'})
+    }
 })
 
 
@@ -88,6 +94,7 @@ router.get('/users/:user_id/workspaces/:workspace_id', function (req, res) {
  * {
  *     workspace_id: workspace id,
  *     workspace_name: workspace name,
+ *     user_role: user's role in the workspace,
  *     date_created: date workspace created,
  *     date_updated: date workspace created
  * }
@@ -95,12 +102,59 @@ router.get('/users/:user_id/workspaces/:workspace_id', function (req, res) {
  * Response Statuses
  * Success - 200 OK
  */
-router.get('/users/:user_id/workspaces', function (req, res) {
-    // TODO - PASS JWT AS user_id FOR AUTHORIZATION???
-    readWorkspaceForUser(req.params.user_id)
-        .then(workspaces => {
-            res.status(200).json(workspaces)
-        })
+router.get('/users/:user_id/workspaces', async function (req, res) {
+    const workspaces = await readWorkspaceForUser(req.params.user_id)
+
+    res.status(200).json(workspaces)
+})
+
+
+/**
+ * Endpoint to get all workspaces that would be candidates for deletion if the user was to be deleted
+ *
+ * Request
+ * Parameter passed via URL path
+ * @param user_id
+ *
+ * Response
+ * @returns workspace candidates for deletion - Array<JSON>
+ * {
+ *     workspace_id: workspace id,
+ *     workspace_name: workspace name,
+ *     user_role: user's role in the workspace,
+ *     date_created: date workspace created,
+ *     date_updated: date workspace created
+ * }
+ *
+ * Response Statuses
+ * Success - 200 OK
+ */
+router.get('/users/:user_id/workspace-candidates-for-deletion', async function (req, res) {
+    const workspaces = await readWorkspaceCandidatesForDeletion(req.params.user_id)
+
+    res.status(200).json(workspaces)
+})
+
+
+/**
+ * FOR TESTING ONLY - Endpoint to get all workspaces
+ *
+ * Response
+ * @returns workspaces - Array<JSON>
+ * {
+ *     workspace_id: workspace id,
+ *     workspace_name: workspace name,
+ *     date_created: date workspace created,
+ *     date_updated: date workspace created
+ * }
+ *
+ * Response Statuses
+ * Success - 200 OK
+ */
+router.get('/workspaces', async function (req, res) {
+    const workspaces = await readAllWorkspaces()
+
+    res.status(200).json(workspaces)
 })
 
 
@@ -118,29 +172,36 @@ router.get('/users/:user_id/workspaces', function (req, res) {
  *
  * Response Statuses
  * Success - 200 OK
- * Failure - 400 Bad Request - Missing a required attribute
+ * Failure - 400 Bad Request
+ * Failure - 403 Forbidden
  * Failure - 404 Not Found
  */
-router.put('/users/:user_id/workspaces/:workspace_id/users/:user_id_to_be_added', function (req, res) {
-    // TODO - AUTH
-    if (req.body.user_role) {
-        addUserToWorkspace(req.params.user_id_to_be_added, req.params.workspace_id, {
-            role_name: req.body.user_role
-        })
-            .then(success => {
-                if (success) {
-                    res.status(200).end()
-                } else {
-                    res.status(404).json({'Error': 'Either the user or workspace ID does not exist'})
-                }
-            })
+router.put('/users/:user_id/workspaces/:workspace_id/users/:user_id_to_be_added', async function (req, res) {
+    const userRole = await readUserRoleInWorkspace(req.params.user_id, req.params.workspace_id)
+
+    if (userRole === 'owner') {
+        if (req.body.user_role) {
+            const userToBeAdded = await readUser(req.params.user_id_to_be_added)
+
+            if (userToBeAdded) {
+                await addUserToWorkspace(req.params.user_id_to_be_added, req.params.workspace_id, {
+                    role_name: req.body.user_role
+                })
+
+                res.status(200).send()
+            } else {
+                res.status(404).json({'Error': 'Either the user or workspace ID does not exist'})
+            }
+        } else {
+            res.status(400).json({
+                    'Error':
+                        'The request object is missing at ' +
+                        'least one of the required attributes',
+                },
+            )
+        }
     } else {
-        res.status(400).json({
-                'Error':
-                    'The request object is missing at ' +
-                    'least one of the required attributes',
-            },
-        )
+        res.status(403).send()
     }
 })
 
@@ -159,29 +220,36 @@ router.put('/users/:user_id/workspaces/:workspace_id/users/:user_id_to_be_added'
  *
  * Response Statuses
  * Success - 200 OK
- * Failure - 400 Bad Request - Missing a required attribute
+ * Failure - 400 Bad Request
+ * Failure - 403 Forbidden
  * Failure - 404 Not Found
  */
-router.patch('/users/:user_id/workspaces/:workspace_id/users/:user_id_to_be_updated', function (req, res) {
-    // TODO - AUTH
-    if (req.body.user_role) {
-        updateUserInWorkspace(req.params.user_id_to_be_updated, req.params.workspace_id, {
-            role_name: req.body.user_role
-        })
-            .then(success => {
-                if (success) {
-                    res.status(200).end()
-                } else {
-                    res.status(404).json({'Error': 'Either the user or workspace ID does not exist'})
-                }
-            })
+router.patch('/users/:user_id/workspaces/:workspace_id/users/:user_id_to_be_updated', async function (req, res) {
+    const userRole = await readUserRoleInWorkspace(req.params.user_id, req.params.workspace_id)
+
+    if (userRole === 'owner') {
+        if (req.body.user_role) {
+            const userToBeUpdated = await readUser(req.params.user_id_to_be_updated)
+
+            if (userToBeUpdated) {
+                await updateUserInWorkspace(req.params.user_id_to_be_updated, req.params.workspace_id, {
+                    role_name: req.body.user_role
+                })
+
+                res.status(200).json(true)
+            } else {
+                res.status(404).json(false)
+            }
+        } else {
+            res.status(400).json({
+                    'Error':
+                        'The request object is missing at ' +
+                        'least one of the required attributes',
+                },
+            )
+        }
     } else {
-        res.status(400).json({
-                'Error':
-                    'The request object is missing at ' +
-                    'least one of the required attributes',
-            },
-        )
+        res.status(403).send()
     }
 })
 
@@ -197,23 +265,30 @@ router.patch('/users/:user_id/workspaces/:workspace_id/users/:user_id_to_be_upda
  *
  * Response Statuses
  * Success - 204 No Content
+ * Failure - 403 Forbidden
  * Failure - 404 Not Found
  */
-router.delete('/users/:user_id/workspaces/:workspace_id/users/:user_id_to_be_deleted', function (req, res) {
-    // TODO - AUTH
-    deleteUserFromWorkspace(req.params.user_id_to_be_deleted, req.params.workspace_id)
-        .then(success => {
-            if (success) {
-                res.status(204).end()
-            } else {
-                res.status(404).json({'Error': 'Either the user or workspace ID does not exist'})
-            }
-        })
+router.delete('/users/:user_id/workspaces/:workspace_id/users/:user_id_to_be_deleted', async function (req, res) {
+    const userRole = await readUserRoleInWorkspace(req.params.user_id, req.params.workspace_id)
+
+    if (userRole === 'owner') {
+        const userToBeDeleted = await readUser(req.params.user_id_to_be_deleted)
+
+        if (userToBeDeleted) {
+            await deleteUserFromWorkspace(req.params.user_id_to_be_deleted, req.params.workspace_id)
+
+            res.status(204).send()
+        } else {
+            res.status(404).json({'Error': 'Either the user or workspace ID does not exist'})
+        }
+    } else {
+        res.status(403).send()
+    }
 })
 
 
 /**
- * Endpoint to edit a workspace
+ * Endpoint to edit a workspace name
  *
  * Request
  * Parameters passed via URL path
@@ -228,23 +303,30 @@ router.delete('/users/:user_id/workspaces/:workspace_id/users/:user_id_to_be_del
  *
  * Response Statuses
  * Success - 200 OK
+ * Failure - 403 Forbidden
  * Failure - 404 Not Found
  */
-router.patch('/users/:user_id/workspaces/:workspace_id', function (req, res) {
-    const workspaceObj = {}
+router.patch('/users/:user_id/workspaces/:workspace_id', async function (req, res) {
+    const userRole = await readUserRoleInWorkspace(req.params.user_id, req.params.workspace_id)
 
-    if (req.body.workspace_name) {
-        workspaceObj.workspace_name = req.body.workspace_name
-    }
+    if (userRole === 'owner') {
+        const workspace = await readWorkspace(req.params.workspace_id)
 
-    updateWorkspace(req.params.workspace_id, workspaceObj)
-        .then(success => {
-            if (success) {
-                res.status(200).json(success)
-            } else {
-                res.status(404).json({'Error': 'No workspace with this workspace id exists'})
+        if (workspace) {
+            const workspaceObj = {}
+
+            if (req.body.workspace_name) {
+                workspaceObj.workspace_name = req.body.workspace_name
             }
-        })
+
+            await updateWorkspace(req.params.workspace_id, workspaceObj)
+            res.status(200).json(true)
+        } else {
+            res.status(404).json(false)
+        }
+    } else {
+        res.status(403).send()
+    }
 })
 
 
@@ -257,18 +339,31 @@ router.patch('/users/:user_id/workspaces/:workspace_id', function (req, res) {
  *
  * Response Statuses
  * Success - 204 No Content
+ * Failure - 403 Forbidden
  * Failure - 404 Not Found
  */
-router.delete('/users/:user_id/workspaces/:workspace_id', function (req, res) {
-    // TODO USE AUTH TO DETERMINE IF USER IS OWNER OF WORKSPACE, PLUS LOGIC TO PASS ALONG OWNERSHIP - MAY REQUIRE PATH CHANGE OR REQUEST BODY
-    deleteWorkspace(req.params.workspace_id)
-        .then(success => {
-            if (success) {
-                res.status(204).end()
-            } else {
-                res.status(404).json({'Error': 'No workspace with this workspace id exists'})
+router.delete('/users/:user_id/workspaces/:workspace_id', async function (req, res) {
+    const workspace = await readWorkspace(req.params.workspace_id)
+
+    if (workspace) {
+        const userRole = await readUserRoleInWorkspace(req.params.user_id, req.params.workspace_id)
+
+        if (userRole === 'owner') {
+            const workspaceProjects = await readProjects(req.params.workspace_id)
+
+            for (const workspaceProject of workspaceProjects) {
+                await deleteProject(workspaceProject.project_id)
             }
-        })
+
+            await deleteWorkspace(req.params.workspace_id)
+
+            res.status(204).send()
+        } else {
+            res.status(403).send()
+        }
+    } else {
+        res.status(404).json({'Error': 'No workspace with this workspace id exists'})
+    }
 })
 
 /* ------------- End Endpoint Functions ------------- */
